requirements.txt
# ── Core data / numerics ──────────────────────────────────────────────────────
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0          # Parquet ingestion (IngestionAgent)

# ── Machine learning ─────────────────────────────────────────────────────────
scikit-learn>=1.3.0      # RandomForest, DummyRegressor/Classifier, LinearModel, SVR, metrics
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0            # Model serialization (BaseModel.save / .load)

# ── Time-series ───────────────────────────────────────────────────────────────
pmdarima>=2.0.0          # ARIMA with auto order selection
statsmodels>=0.14.0      # pmdarima dependency; also used for ARIMA internals

# ── Deep learning ─────────────────────────────────────────────────────────────
torch>=2.1.0             # LSTM model + Chronos runtime

# ── Chronos (Amazon pretrained forecaster) ────────────────────────────────────
chronos-forecasting>=1.3.0

# ── Hyperparameter optimisation ───────────────────────────────────────────────
optuna>=3.4.0

# ── LLM API (IngestionAgent + ProblemClassificationAgent) ────────────────────
openai>=1.30.0

# ── Plotting (EvaluationAgent, ArtifactAssemblyAgent) ────────────────────────
matplotlib>=3.8.0

# ── Schema validation ─────────────────────────────────────────────────────────
jsonschema>=4.21.0       # LLM response validation in IngestionAgent

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
